Move test server context to fixtures
