+++ application/src/test/java/org/opentripplanner/apis/vectortiles/DebugVectorTilesResourceTest.java
@@ -4,10 +4,10 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.List;
 import org.junit.jupiter.api.Test;
-import org.opentripplanner.TestServerContext;
 import org.opentripplanner.apis.vectortiles.model.LayerType;
 import org.opentripplanner.ext.fares.service.NoopFareServiceFactory;
 import org.opentripplanner.inspector.vector.LayerParameters;
+import org.opentripplanner.standalone.api.TestServerContext;
 import org.opentripplanner.street.graph.Graph;
 import org.opentripplanner.test.support.HttpForTest;
 import org.opentripplanner.transfer.regular.TransferServiceTestFactory;
